added toggle jsx for better organization
